chore: unify integrated tests
Prior to this commit, the integrated tests were executed on serial
different `tokio::tests` instances. This caused the need to serialize
and deserialize the public parameters to prevent them from being
generated for each run.

Such serialization caused issues on the CI, where multiple instances of
the tests were executed in parallel, causing a huge load on the disk
storage.

This commit unifies the integrated tests execution into a single
function. It will not only better emulate a real production environment,
but will also unlock the possibility to generate the public parameters
only once and never serialize it.