modify groth verifier contracts for better test-ability and extendability (#414)
This PR does the following:
* adds the `virtual` tag to several functions in
`Groth16VerifierExtension.sol` - this allows us to override the
functions in a parent contract, which makes it much easier to test. With
these changes, we can test as much of the response codepath as possible
without actually having to generate proofs. This will also make it
easier to tweak or extend verification logic for specific chains. Ex on
polygon (and in the future on scroll) we want to skip blockhash
verification. This will be much easier to do once we have this setup.
* moves `verifyBlockhash()` to it's own function. This is specifically
for the `isCDK()` case where we want to override this function.
* renames `verifier.sol` => `Verifier.sol` and
`Groth16VerifierExtensions.sol` => `Groth16VerifierExtension.sol` for
consistency with the contracts codebase
* Changes the name of the `Query` contract to `Groth16VerifierExtension`
for consistency