Use Batching Query Public Inputs for Tabular Queries + Remove Old Circuits (#417)
This PR cleans up the circuit code in `verifiable-db` crate. In
particular

- Rather than keeping 2 versions of query circuits (batched and
non-batched) in the codebase, the old non-batched circuits, which will
become useless when batching circuits will be integrated, are removed
- To avoid having to deal with 2 different set of public inputs in query
circuits API, the main change of this PR is making tabular queries
circuits compatible with the set of public inputs exposed by batching
circuits. This will yield some unnecessary public inputs for tabular
queries, but it will allow to keep a single API to generate query proofs
- As a consequence of the removal of the old non-batched circuits, also
the integration test code for queries can be greatly simplified

---------

Co-authored-by: T <asongala@163.com>