Deterministic `PublicParameters` Serialisation (#437)
This PR resolves CRY-35 and is joint work with @nicholas-mainardi. 

It changes the use of a `HashMap<Vec<F>, usize>` in
`recursion-framework/src/universal_verifier_gadget` to a
`BTreeMap<Vec<u64>, usize>` for deterministic iteration.

In addition to this it incorporates @nicholas-mainardi solution for
avoiding directly serialising the `empty_node_proof` in
`verifiable_db::cells_tree::PublicParamaters`.