feat: provable commitment for off chain tables (#477)
This PR adds the possibility to use a provable commitment to the data
inserted in an off-chain table as a root of trust for such a table. It
adds also a public API to re-compute the same commitment computed in
circuits, which can be used to check that the zkTable has been built
with the same data bound to the commitment.